Various updates listed below.
Separate Clause into Rule or Fact in AST.
Make EDB Predicate indirectly held by a Literal for symmetry in the use
of RAGs.
Coalesce predicates that refer to the same relation into a
"SuperPredicate". A SuperPredicate is an NTA node with a Set of
predicate objects that it refers to through an attribute. This greatly
simplifies various error checks.

Added various attributes to the AST nodes.
In particular back-pointer hierarchy: Predicate-->Literal-->Stmt-->Program
A predicate is either used in a Rule or in a Fact.
==> The use in a rule is either a USE or a DEF.
==> The use in a fact is either a FILE fact or a EDB fact.

Added Arity Compution and Check.
Added Error if use Predicate is always empty (maybe should change to
        warning)

Added Additional Tests.