Refactor the doEvaluationTest method to use doSingleEvaluation
This uncovered a bug in the original doEvaluationTest, where the
only the OUTPUT formal predicate for program1 was used. With this
fixed, the withmeta/evalTest_5 started to fail and was disables.

I suspect that this happens because there is a bug in pupulating
the OUTPUT formal predicate, in cases like:
EDB('OUTPUT, "OUTPUT.csv").
A(x) :- OUTPUT(x).