Change the syntax of the analyze blocks
The analyze blocks now load the facts from the analyzed program
and expose it through patterns and context predicates. The argument
to the analyze block is a unary predicate containing the paths
to be searched for source files.

This removes the need of using the IMPORT predicate, which is now
obsolete, and hides the program representation relation from the user.
This allows us more freedom into how we represent different
aspects of a program: AST, attributes, provenance information, etc.