Restructure the internal datatabase
The internal database contained a set of tables for each compilation
unit. This added a huge number of tables to the DB (~ 10000 tables),
which made the reads and writes very slow.

Replace this structure with one table for each relation. Inside the
table, each tuple is tagged with the ID of the file that produced
the tuple.